+++ src/routes/utils/icons/Clipboard.svelte
@@ -1,0 +1,8 @@
+<svg width="14" height="18" viewBox="0 0 14 18" fill="none" xmlns="http://www.w3.org/2000/svg">
+  <path
+    d="M4.49984 3.16667H2.83317C2.39114 3.16667 1.96722 3.34226 1.65466 3.65482C1.3421 3.96738 1.1665 4.39131 1.1665 4.83333V14.8333C1.1665 15.2754 1.3421 15.6993 1.65466 16.0118C1.96722 16.3244 2.39114 16.5 2.83317 16.5H11.1665C11.6085 16.5 12.0325 16.3244 12.345 16.0118C12.6576 15.6993 12.8332 15.2754 12.8332 14.8333V4.83333C12.8332 4.39131 12.6576 3.96738 12.345 3.65482C12.0325 3.34226 11.6085 3.16667 11.1665 3.16667H9.49984M4.49984 3.16667C4.49984 3.60869 4.67543 4.03262 4.98799 4.34518C5.30055 4.65774 5.72448 4.83333 6.1665 4.83333H7.83317C8.2752 4.83333 8.69912 4.65774 9.01168 4.34518C9.32424 4.03262 9.49984 3.60869 9.49984 3.16667M4.49984 3.16667C4.49984 2.72464 4.67543 2.30072 4.98799 1.98816C5.30055 1.67559 5.72448 1.5 6.1665 1.5H7.83317C8.2752 1.5 8.69912 1.67559 9.01168 1.98816C9.32424 2.30072 9.49984 2.72464 9.49984 3.16667"
+    stroke="currentColor"
+    stroke-width="2"
+    stroke-linecap="round"
+    stroke-linejoin="round" />
+</svg>
